requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.5
argparse==1.2.1
itsdangerous==0.24
pyserial==3.0.1
wsgiref==0.1.2
